requirements.txt
asgiref==3.2.7
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
gunicorn==20.0.4
dj-database-url
markdown==3.0.0
opencv-python==4.2.0.32
numpy==1.18.2
imutils==0.5.3
jsonfield==3.1.0